requirements.txt
aiogram==3.5.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pillow==9.5.0
httpx==0.27.0
anyio==4.3.0pip install --no-cache-dir --no-deps -r requirements-fixed.txt